spikes/weekend_agent/pyproject.toml
[tool.poetry]
name = "weekend-agent"
version = "0.1.0"
description = ""
authors = ["Rajesh Korde <rajesh.korde@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.64"
langchain-community = "^0.3.14"
langchain-openai = "^0.3.0"
python-dotenv = "^1.0.1"
tavily-python = "^0.5.0"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
